+++ repo/common_repo.go
@@ -2,6 +2,9 @@ package commonrepo
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"reflect"
 	"time"
 
 	"github.com/sirupsen/logrus"
@@ -9,6 +12,8 @@ import (
 	"go.mongodb.org/mongo-driver/bson"
 
 	"go.mongodb.org/mongo-driver/mongo"
+
+	mongopagination "github.com/gobeam/mongo-go-pagination"
 )
 
 type IDBOperation interface {
@@ -16,10 +21,24 @@ type IDBOperation interface {
 	FindOne(filter, impl interface{}) (interface{}, error)
 	FindOneAndUpdate(filter, data interface{}) (*mongo.UpdateResult, error)
 	Find(filter, impl interface{}) (interface{}, error)
+	FindPagedSorted(pagingSortingReq PagingSortingRequest, filter, impl interface{}) (interface{}, *mongopagination.PaginationData, error)
 	InsertOneAtColl(collection string, data interface{}) (*mongo.InsertOneResult, error)
 	FindOneAtColl(collection string, filter, impl interface{}) (interface{}, error)
 	FindOneAndUpdateAtColl(collection string, filter, data interface{}) (*mongo.UpdateResult, error)
 	FindAtColl(collection string, filter, impl interface{}) (interface{}, error)
+	FindAtCollPagedSorted(collection string, pagingSortingReq PagingSortingRequest, filter, impl interface{}) (interface{}, *mongopagination.PaginationData, error)
+}
+
+//PagingSortingRequest specify paging and sorting mechanism
+//SortByField fill with name of the field to be taken by sorter
+//IsSortDesc leave empty (default false) if sort in ascending order
+//LimitPerPage number of items per page
+//Page page index (start at 1)
+type PagingSortingRequest struct {
+	SortByField  string
+	IsSortDesc   bool
+	LimitPerPage int64
+	Page         int64
 }
 
 type DBOperation struct {
@@ -50,6 +69,10 @@ func (dbOp *DBOperation) FindOneAndUpdate(filter, data interface{}) (*mongo.Upda
 
 func (dbOp *DBOperation) Find(filter, impl interface{}) (interface{}, error) {
 	return dbOp.FindAtColl(dbOp.defaultCollection, filter, impl)
+}
+
+func (dbOp *DBOperation) FindPagedSorted(pagingSortingReq PagingSortingRequest, filter, impl interface{}) (interface{}, *mongopagination.PaginationData, error) {
+	return dbOp.FindAtCollPagedSorted(dbOp.defaultCollection, pagingSortingReq, filter, impl)
 }
 
 func (dbOp *DBOperation) InsertOneAtColl(collection string, data interface{}) (*mongo.InsertOneResult, error) {
@@ -119,4 +142,71 @@ func (dbOp *DBOperation) FindAtColl(collection string, filter, impl interface{})
 	}
 
 	return impl, nil
+}
+
+//FindAtCollPagedSorted find at specified collection and return as paged and sorted result
+func (dbOp *DBOperation) FindAtCollPagedSorted(collection string, pagingSortingReq PagingSortingRequest, filter, impl interface{}) (interface{}, *mongopagination.PaginationData, error) {
+	coll := dbOp.dbClient.Database(dbOp.databaseName).Collection(collection)
+	paginatedData, err := mongopagination.New(coll).
+		Limit(pagingSortingReq.LimitPerPage).
+		Page(pagingSortingReq.Page).
+		Sort(pagingSortingReq.SortByField, dbOp.convertToSortType(pagingSortingReq.IsSortDesc)).
+		Filter(filter).
+		Find()
+
+	if err != nil {
+		logrus.Errorf("Failed to perform query with sort at coll %s. PagingSortingReq: %v | Error: %v", collection, pagingSortingReq, err)
+		return nil, nil, err
+	}
+
+	if paginatedData.Data == nil {
+		return nil, nil, nil
+	}
+
+	err = dbOp.decodeBsonRaws(paginatedData.Data, impl)
+	if err != nil {
+		logrus.Errorf("Failed to perform dbOp.decodeBsonRaws(paginatedData.Data, impl). Impl: %v | Error: %v", impl, err)
+		return nil, nil, err
+	}
+
+	return impl, &paginatedData.Pagination, nil
+}
+
+//decodeBsonRaws iterate the bson.Raw array and decode to impl
+func (dbOp *DBOperation) decodeBsonRaws(rawData []bson.Raw, impl interface{}) error {
+	resultsVal := reflect.ValueOf(impl)
+	if resultsVal.Kind() != reflect.Ptr {
+		return fmt.Errorf("non-pointer %v", resultsVal.Type())
+	}
+	// get the value that the pointer v points to.
+	sliceVal := resultsVal.Elem()
+	if sliceVal.Kind() != reflect.Slice {
+		return errors.New("can't fill non-slice value")
+	}
+	elemType := sliceVal.Type().Elem()
+	var index int
+	for _, w := range rawData {
+		if sliceVal.Len() == index {
+			// slice is full
+			newElem := reflect.New(elemType)
+			sliceVal = reflect.Append(sliceVal, newElem.Elem())
+			sliceVal = sliceVal.Slice(0, sliceVal.Cap())
+		}
+
+		currElem := sliceVal.Index(index).Addr().Interface()
+		err := bson.Unmarshal(w, currElem)
+		if err != nil {
+			return fmt.Errorf("can't unmarshal. Err: %v", err)
+		}
+		index++
+	}
+	resultsVal.Elem().Set(sliceVal.Slice(0, index))
+	return nil
+}
+
+func (dbOp *DBOperation) convertToSortType(sortDesc bool) int {
+	if sortDesc {
+		return -1
+	}
+	return 1
 }